backend/requirements.txt
FastAPI
uvicorn
pydantic
azure-storage-file-datalake
boto3
smbprotocolnode -v
npm -v
schedule
cryptography